+++ configs/embed.go
@@ -1,0 +1,6 @@
+package configs
+
+import _ "embed"
+
+//go:embed default.yaml
+var Default string