+++ internal/installer/network.go
@@ -70,7 +70,7 @@ func (n *NetworkChecker) checkHost(ctx context.Context, url string) error {
 	if err != nil {
 		return err
 	}
-	defer resp.Body.Close()
+	defer func() { _ = resp.Body.Close() }()
 
 	// Any response means we have connectivity
 	return nil